feature pengguna last

+++ app/Http/Controllers/AdminCms/User.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\AdminCms;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -12,39 +12,40 @@ class User extends Controller
     public function index()
     {
     	if(Session()->get('username')=="") { return redirect('login')->with(['warning' => 'Mohon maaf, Anda belum login']);}
-		$user 	= DB::table('users')->orderBy('id_user','DESC')->get();
-
-		$data = array(  'title'     => 'Pengguna Website',
+		$user 	= DB::connection('ts3')->table('auth.v_list_user')->orderBy('id_user','DESC')->get();
+        $role 	= DB::connection('ts3')->table('auth.user_roles')->where('id', '!=' , 1)->get();
+		$data = array(  'title'     => 'Pengguna Sistem',
 						'user'      => $user,
-                        'content'   => 'admin/user/index'
+                        'roledata'      => $role,
+                        'content'   => 'admin-cms/user/index'
                     );
-        return view('admin/layout/wrapper',$data);
+        return view('admin-cms/layout/wrapper',$data);
     }
 
     // Index
     public function edit($id_user)
     {
         if(Session()->get('username')=="") { return redirect('login')->with(['warning' => 'Mohon maaf, Anda belum login']);}
-        $user   = DB::table('users')->where('id_user',$id_user)->orderBy('id_user','DESC')->first();
+        $user   =DB::connection('ts3')->table('auth.users')->where('id_user',$id_user)->orderBy('id_user','DESC')->first();
 
         $data = array(  'title'     => 'Edit Pengguna Website',
                         'user'      => $user,
-                        'content'   => 'admin/user/edit'
+                        'content'   => 'admin-cms/user/edit'
                     );
-        return view('admin/layout/wrapper',$data);
+        return view('admin-cms/layout/wrapper',$data);
     }
 
     // Proses
     public function proses(Request $request)
     {
-        $site   = DB::table('konfigurasi')->first();
+        $site   =DB::connection('ts3')->table('konfigurasi')->first();
         // PROSES HAPUS MULTIPLE
         if(isset($_POST['hapus'])) {
             $id_usernya       = $request->id_user;
             for($i=0; $i < sizeof($id_usernya);$i++) {
-                DB::table('users')->where('id_user',$id_usernya[$i])->delete();
+               DB::connection('ts3')->table('auth.users')->where('id_user',$id_usernya[$i])->delete();
             }
-            return redirect('admin/user')->with(['sukses' => 'Data telah dihapus']);
+            return redirect('admin-cms/user')->with(['sukses' => 'Data telah dihapus']);
         // PROSES SETTING DRAFT
         }
     }
@@ -55,7 +56,7 @@ class User extends Controller
     	if(Session()->get('username')=="") { return redirect('login')->with(['warning' => 'Mohon maaf, Anda belum login']);}
     	request()->validate([
                             'nama'     => 'required',
-					        'username' => 'required|unique:users',
+					        'username' => 'required',
 					        'password' => 'required',
                             'email'    => 'required',
                             'gambar'   => 'file|image|mimes:jpeg,png,jpg|max:8024',
@@ -76,24 +77,31 @@ class User extends Controller
             $destinationPath = './assets/upload/user/';
             $image->move($destinationPath, $input['nama_file']);
             // END UPLOAD
-            DB::table('users')->insert([
+            
+           DB::connection('ts3')->table('auth.users')->insert([
                 'nama'          => $request->nama,
                 'email'	        => $request->email,
                 'username'   	=> $request->username,
                 'password'      => sha1($request->password),
-                'akses_level'   => $request->akses_level,
-                'gambar'        => $input['nama_file']
+                'id_role'       => $request->role,
+                'gambar'        => $input['nama_file'],
+                'created_at'    => date("Y-m-d h:i:sa"),
+                'create_by'     => $request->session()->get('username'),
+                'is_active'     => true
             ]);
         }else{
-             DB::table('users')->insert([
+            DB::connection('ts3')->table('auth.users')->insert([
                 'nama'          => $request->nama,
                 'email'         => $request->email,
                 'username'      => $request->username,
                 'password'      => sha1($request->password),
-                'akses_level'   => $request->akses_level
+                'id_role'       => $request->role,
+                'created_at'    => date("Y-m-d h:i:sa"),
+                'create_by'     => $request->session()->get('username'),
+                'is_active'     => true
             ]);
         }
-        return redirect('admin/user')->with(['sukses' => 'Data telah ditambah']);
+        return redirect('admin-cms/user')->with(['sukses' => 'Data telah ditambah']);
     }
 
     // edit
@@ -125,32 +133,32 @@ class User extends Controller
             $image->move($destinationPath, $input['nama_file']);
             // END UPLOAD
             $slug_user = Str::slug($request->nama, '-');
-            DB::table('users')->where('id_user',$request->id_user)->update([
+           DB::connection('ts3')->table('auth.users')->where('id_user',$request->id_user)->update([
                 'nama'          => $request->nama,
                 'email'         => $request->email,
                 'username'      => $request->username,
                 'password'      => sha1($request->password),
-                'akses_level'   => $request->akses_level,
+                'id_role'   => $request->id_role,
                 'gambar'        => $input['nama_file']
             ]);
         }else{
             $slug_user = Str::slug($request->nama, '-');
-            DB::table('users')->where('id_user',$request->id_user)->update([
+           DB::connection('ts3')->table('auth.users')->where('id_user',$request->id_user)->update([
                 'nama'          => $request->nama,
                 'email'         => $request->email,
                 'username'      => $request->username,
                 'password'      => sha1($request->password),
-                'akses_level'   => $request->akses_level
+                'id_role'   => $request->id_role
             ]);
         }
-        return redirect('admin/user')->with(['sukses' => 'Data telah diupdate']);
+        return redirect('admin-cms/user')->with(['sukses' => 'Data telah diupdate']);
     }
 
     // Delete
     public function delete($id_user)
     {
     	if(Session()->get('username')=="") { return redirect('login')->with(['warning' => 'Mohon maaf, Anda belum login']);}
-    	DB::table('users')->where('id_user',$id_user)->delete();
-    	return redirect('admin/user')->with(['sukses' => 'Data telah dihapus']);
+    	DB::connection('ts3')->table('auth.users')->where('id_user',$id_user)->delete();
+    	return redirect('admin-cms/user')->with(['sukses' => 'Data telah dihapus']);
     }
 }
